block win_walue repeated

+++ ERA/src/controller/Controller.java
@@ -2369,7 +2369,7 @@ public class Controller extends Observable {
 
 		/// return this.blockChain != null?
 		/// this.blockChain.getHWeightFull(dcSet).a: -1;
-		Tuple2<Integer, Long> hWeight = dcSet.getBlockSignsMap().get(blockChain.getLastBlockSignature(dcSet));
+		Tuple2<Integer, Integer> hWeight = dcSet.getBlockSignsMap().get(blockChain.getLastBlockSignature(dcSet));
 		if (hWeight == null || hWeight.a == -1)
 			return -1;
